refactor: move router, theme to index
