+++ src/constants/theme-default.ts
@@ -1,0 +1,15 @@
+import { DefaultTheme } from 'styled-components';
+
+const theme: DefaultTheme = {
+  fg: '#ffe190',
+  fgError: '#d4eee8',
+  bg: '#237773',
+  bgActive: '#6fb38e',
+  scrollbarBorderRadius: '8px',
+  scrollbarWidth: '10px',
+  scrollbarBg: '#6fb38e',
+  scrollbarThumbWidth: '8px',
+  scrollbarThumbBg: '#1d3e3c',
+};
+
+export default theme;
