+++ src/index.js
@@ -1,15 +1,22 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import * as serviceWorker from './serviceWorker';
 import App from './components/app';
 import store from './redux/store';
 import './index.css';
+import { ThemeProvider } from 'styled-components';
+import theme from './constants/theme-default';
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <BrowserRouter>
+        <ThemeProvider theme={theme}>
+          <App/>
+        </ThemeProvider>
+      </BrowserRouter>
     </Provider>
   </React.StrictMode>,
   document.getElementById('root')
